refactor: rename api view files to genereic view name
