+++ infoenergia_api/api/contracts/__init__.py
@@ -1,1 +1,1 @@
-from .contracts import bp_contracts
+from .views import bp_contracts
