+++ infoenergia_api/api/f1_measures/__init__.py
@@ -1,1 +1,1 @@
-from .f1_measures import bp_f1_measures
+from .views import bp_f1_measures
