+++ infoenergia_api/api/tariff/__init__.py
@@ -1,1 +1,1 @@
-from .tariff import bp_tariff
+from .views import bp_tariff
